fixed landing.css (#43)
* added styles counters

* fixed bug in price in my assets, added tokenId in the name, added filter by tokenId

* Fix nft spec

---------

+++ app/graphql/types/nft_type.rb
@@ -4,7 +4,7 @@ module Types
   class NftType < Types::BaseObject
     field :id, ID, null: false
     field :price, Float
-    field :tokenId, Integer
+    field :tokenId, Float
     field :seller, String
     field :owner, String
     field :image, String
